removed addButtonListener fn and added onclick="getDetails('${hero.name}') directly to code creating each button
This way the addButtonListener fn isn't called for each existing character every time a new character is loaded. It had the effect of adding 200 event listeners to the first character loaded if 199 more loaded afterwards, which was causing an approx 1.5 second delay when button clicked waiting for modal to open. Now ALL modals open quickly , not just last loaded characters.